fix(api): delegate minors lookup to service
- Move minors_index query to minorIndexService
- Keep /usuarios/[uid]/menores route thin and service-driven
- Standardize minors response mapping
